requirements.txt
fastapi
uvicorn[standard]
sqlalchemy
pydantic
python-dotenv
pytest
pymysql
sqlmodel
httpx
boto3
twilio
os